+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machines/dna_decryptor/DnaDecryptorBlockEntity.kt
@@ -1,7 +1,7 @@
 package dev.aaronhowser.mods.geneticsresequenced.block.machines.dna_decryptor
 
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.Gene
-import dev.aaronhowser.mods.geneticsresequenced.api.genes.MobGeneRegistry
+import dev.aaronhowser.mods.geneticsresequenced.data.MobGeneRegistry
 import dev.aaronhowser.mods.geneticsresequenced.block.base.CraftingMachineBlockEntity
 import dev.aaronhowser.mods.geneticsresequenced.gene.ModGenes
 import dev.aaronhowser.mods.geneticsresequenced.item.DnaHelixItem